% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchRefuginator.R
\name{launchRefuginator}
\alias{launchRefuginator}
\title{Launch the Refuginator Application in your Browser}
\usage{
launchRefuginator(defaultbrowser = TRUE)
}
\arguments{
\item{defaultbrowser}{Choose whether to launch the application in your default browser.}
}
\value{
Launches Refuginator application in your browser. All features are contained within the application. Read the "refuginatorIntro" vignette for a guide to the features.
}
\description{
Launch the Refuginator Application in your Browser
}
\examples{
\dontrun{launchRefuginator()}

}
