% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchRefuginator.R
\name{launchRefuginator}
\alias{launchRefuginator}
\title{Launch the Refuginator Application in your Browser}
\usage{
launchRefuginator(inbrowser = TRUE)
}
\arguments{
\item{inbrowser}{Choose whether to launch the application in your default browser.}
}
\value{
Launches Refuginator application in your browser. All features are contained within the application.
}
\description{
Launch the Refuginator Application in your Browser
}
\examples{
launchRefuginator()
launchRefuginator(inbrowser = TRUE)
}
